+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `UsbMidiEventPacket::to_raw_bytes` function.
 - `UsbMidiEventPacket::from_message_bytes` function.
 - `TryFrom<&UsbMidiEventPacket>` implementation for `Message` type.
+- `Message::into_packet` function.
 - Re-exports of common items in crate root.
 
 ### Changed